requirements.txt
Django==4.2.20
djangorestframework==3.15.2
django-cors-headers==4.3.1

# Authentication

django-allauth==0.60.1
djangorestframework-simplejwt==5.3.1
argon2-cffi==23.1.0

# Database

psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Environment variables

python-dotenv==1.0.0

# Image Processing

Pillow==10.3.0

# API Integration

requests==2.32.0
httpx==0.26.0

# Money handling

django-money==3.4.1

# Payment Processing

stripe==7.10.0

# File Upload

django-storages==1.14.2

# Cache

django-redis==5.4.0

# Development

django-debug-toolbar==4.2.0
black==24.3.0
flake8==7.0.0
isort==5.13.2

# Testing

pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0

# Production

gunicorn==22.0.0
whitenoise==6.6.0

# Task Queue

celery==5.3.6
redis==5.0.1

# Pokemon TCG SDK
pokemontcgsdk==3.4.0
